requirements.txt
# core
Cython>=0.27,<1
Jinja2>=2,<3
pyenet
toml
Pillow>=5.1.0,<7
aiohttp>=3.3.0,<3.7.0
packaging>=19.0

# from command
geoip2>=2.9,<3.0

# [tls] for core, [conch] for ssh
Twisted[tls,conch]

# windows specific
pypiwin32; platform_system=="Windows"
